% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\name{clear}
\alias{clear}
\title{Clears all completed tasks from the specified task list. The affected tasks will be marked as 'hidden' and no longer be returned by default when retrieving all tasks for a task list.}
\usage{
clear(tasklist)
}
\arguments{
\item{tasklist}{Task list identifier}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tasks)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/tasks/firstapp}{Google Documentation}
}

