% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\name{insert}
\alias{insert}
\title{Creates a new task on the specified task list.}
\usage{
insert(Task, tasklist, parent = NULL, previous = NULL)
}
\arguments{
\item{Task}{The \link{Task} object to pass to this method}

\item{tasklist}{Task list identifier}

\item{parent}{Parent task identifier}

\item{previous}{Previous sibling task identifier}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tasks)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/tasks/firstapp}{Google Documentation}

Other Task functions: \code{\link{Task.links}},
  \code{\link{Task}}, \code{\link{patch}},
  \code{\link{update}}
}
