% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\name{move}
\alias{move}
\title{Moves the specified task to another position in the task list. This can include putting it as a child task under a new parent and/or move it to a different position among its sibling tasks.}
\usage{
move(tasklist, task, parent = NULL, previous = NULL)
}
\arguments{
\item{tasklist}{Task list identifier}

\item{task}{Task identifier}

\item{parent}{New parent task identifier}

\item{previous}{New previous sibling task identifier}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tasks)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/tasks/firstapp}{Google Documentation}
}
