% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_objects.R
\name{Task.links}
\alias{Task.links}
\title{Task.links Object}
\usage{
Task.links()
}
\value{
Task.links object
}
\description{
Task.links Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Task functions: \code{\link{Task}},
  \code{\link{insert}}, \code{\link{patch}},
  \code{\link{update}}
}
