% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_objects.R
\name{TaskList}
\alias{TaskList}
\title{TaskList Object}
\usage{
TaskList(etag = NULL, id = NULL, selfLink = NULL, title = NULL,
  updated = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{id}{Task list identifier}

\item{selfLink}{URL pointing to this task list}

\item{title}{Title of the task list}

\item{updated}{Last modification time of the task list (as a RFC 3339 timestamp)}
}
\value{
TaskList object
}
\description{
TaskList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TaskList functions: \code{\link{tasklists.insert}},
  \code{\link{tasklists.patch}},
  \code{\link{tasklists.update}}
}

