% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\name{tasklists.insert}
\alias{tasklists.insert}
\title{Creates a new task list and adds it to the authenticated user's task lists.}
\usage{
tasklists.insert(TaskList)
}
\arguments{
\item{TaskList}{The \link{TaskList} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tasks)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/tasks/firstapp}{Google Documentation}

Other TaskList functions: \code{\link{TaskList}},
  \code{\link{tasklists.patch}},
  \code{\link{tasklists.update}}
}

