% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\name{tasklists.list}
\alias{tasklists.list}
\title{Returns all the authenticated user's task lists.}
\usage{
tasklists.list(maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of task lists returned on one page}

\item{pageToken}{Token specifying the result page to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
\item https://www.googleapis.com/auth/tasks.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tasks, https://www.googleapis.com/auth/tasks.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/tasks/firstapp}{Google Documentation}
}

