% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_objects.R
\name{TaskLists}
\alias{TaskLists}
\title{TaskLists Object}
\usage{
TaskLists(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{Collection of task lists}

\item{nextPageToken}{Token that can be used to request the next page of this result}
}
\value{
TaskLists object
}
\description{
TaskLists Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
