% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\name{tasklists.update}
\alias{tasklists.update}
\title{Updates the authenticated user's specified task list.}
\usage{
tasklists.update(TaskList, tasklist)
}
\arguments{
\item{TaskList}{The \link{TaskList} object to pass to this method}

\item{tasklist}{Task list identifier}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tasks)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/tasks/firstapp}{Google Documentation}

Other TaskList functions: \code{\link{TaskList}},
  \code{\link{tasklists.insert}},
  \code{\link{tasklists.patch}}
}

