% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_objects.R
\name{Tasks}
\alias{Tasks}
\title{Tasks Object}
\usage{
Tasks(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{items}{Collection of tasks}

\item{nextPageToken}{Token used to access the next page of this result}
}
\value{
Tasks object
}
\description{
Tasks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
