% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_functions.R
\docType{package}
\name{tasks_googleAuthR}
\alias{tasks_googleAuthR}
\alias{tasks_googleAuthR-package}
\title{Tasks API
Lets you manage your tasks and task lists.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:49:39
filename: /Users/mark/dev/R/autoGoogleAPI/googletasksv1.auto/R/tasks_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/tasks
\item https://www.googleapis.com/auth/tasks.readonly
}
}

